requirements.txt
# Core dependencies
PyMuPDF>=1.20.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.3
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
datasketch>=1.5.0

# FastAPI backend
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.5
pydantic>=2.0.0
pydantic-settings>=2.0.1
tenacity>=8.2.0

# PDF processing
pdf2image>=1.16.0
pillow>=9.0.0
pytesseract>=0.3.10
pypdf>=3.7.0
python-docx>=0.8.11

# Utilities
python-dotenv>=1.0.0
tabulate>=0.9.0
tqdm>=4.66.0
colorama>=0.4.6

# Testing and development
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
pytest-cov>=4.1.0